cognita-v2-main/backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
groq==0.9.0
PyPDF2==3.0.1
python-dotenv==1.0.0
anyio==3.7.1
httpx==0.27.0
